feat: full audit remediation + Paper 2 math integration + embedded Opus reasoner
Audit Remediation (Steps 1-7):
- Fix repository artifact integrity: track spine, backfill, v1_views modules
- Make PMM2 activation explicit and fail-closed on init errors
- Centralize live mutation safety via MutationGuard in OrderManager
- Gate quote resume on reconciliation freshness token
- Establish canonical fill/fee contract with exchange identity dedup
- Unify exposure semantics to mark-to-market dollars for risk limits
- Bring whole-repo verification to green: 0 ruff, 0 mypy strict, 287 tests

Paper 2 Mathematical Integration:
- pmm1/math/kelly.py: Kelly criterion with fractional sizing, growth rate
- pmm1/math/extremize.py: log-odds extremization to fix RLHF hedging
- pmm1/math/ensemble.py: log-odds pooling, MWU, BMA, inverse-Brier weights
- pmm1/math/validation.py: SPRT, power analysis, rolling Sharpe, Brier, ECE
- Kelly sizing in QuoteEngine (behind kelly_enabled flag)
- EdgeTracker with SPRT + dynamic edge confidence feeding Kelly fraction
- FairValueCalibrator tracking Brier/ECE on accumulated samples
- Absolute 15% cumulative drawdown kill switch
- V3 calibration upgraded with extremization + log-odds ensemble pooling
- MWU model weight tracker for online multi-model adaptation
- V3 integrator upgraded with Paper 1 weighted blend (67/33 market/AI)

Embedded Opus Reasoner:
- pmm1/strategy/llm_reasoner.py: background Opus 4.6 reasoning loop
- OAuth token auth (personal subscription, no API billing)
- Extended thinking with prompt caching
- Non-blocking: background loop writes cache, quote loop reads
- Signal decay toward market midpoint with configurable half-life
- Activated via PMM1_LLM_ENABLED + ANTHROPIC_OAUTH_TOKEN env vars

All new features deploy dormant by default (zero behavior change).

+++ pmm2/scorer/rebate_ev.py
@@ -46,7 +46,7 @@ def compute_rebate_ev(
     # feeEq = size * price * fee_rate * (price * (1 - price)) ^ exp_m
     def fee_eq(size: float, price: float) -> float:
         variance = price * (1.0 - price)
-        return size * price * market.fee_rate * (variance**exp_m)
+        return float(size * price * market.fee_rate * (variance**exp_m))
 
     our_fee_eq_bid = fee_eq(bundle.bid_size, bundle.bid_price)
     our_fee_eq_ask = fee_eq(bundle.ask_size, bundle.ask_price)
